Utils and refactor escape-label
