+++ packages/cypher-builder/src/clauses/sub-clauses/Projection.ts
@@ -19,8 +19,9 @@
 
 import type { CypherEnvironment } from "../../Environment";
 import type { Expr } from "../../types";
-import type { Literal, Variable } from "../../CypherBuilder";
 import { CypherASTNode } from "../../CypherASTNode";
+import type { Variable } from "../../variables/Variable";
+import type { Literal } from "../../variables/Literal";
 
 // TODO: improve projection column, some expressions require alias
 export type ProjectionColumn = Expr | [Expr, string | Variable | Literal];
