+++ packages/cypher-builder/src/clauses/sub-clauses/Set.ts
@@ -18,12 +18,12 @@
  */
 
 import { CypherASTNode } from "../../CypherASTNode";
-import type { MapProjection } from "../../CypherBuilder";
 import type { CypherEnvironment } from "../../Environment";
 import type { MapExpr } from "../../expressions/map/MapExpr";
 import type { PropertyRef } from "../../variables/PropertyRef";
 import type { Expr } from "../../types";
 import { padBlock } from "../../utils/utils";
+import type { MapProjection } from "../../expressions/map/MapProjection";
 
 export type SetParam = [PropertyRef, Exclude<Expr, MapExpr | MapProjection>];
