+++ packages/cypher-builder/src/expressions/HasLabel.ts
@@ -19,7 +19,7 @@
 
 import { CypherASTNode } from "../CypherASTNode";
 import type { CypherEnvironment } from "../Environment";
-import { escapeLabel } from "../utils/utils";
+import { escapeLabel } from "../utils/escape-label";
 import type { NodeRef } from "../variables/NodeRef";
 
 export class HasLabel extends CypherASTNode {
