+++ packages/cypher-builder/src/index.ts
@@ -103,3 +103,10 @@ export type { SetParam } from "./clauses/sub-clauses/Set";
 export type { PredicateFunction } from "./expressions/functions/PredicateFunctions";
 export type { Order } from "./clauses/sub-clauses/OrderBy";
 export type { Pattern } from "./Pattern";
+
+// utils
+import { escapeLabel } from "./utils/escape-label";
+
+export const Utils = {
+    escapeLabel,
+};
