+++ packages/cypher-builder/tests/utils.test.ts
@@ -17,22 +17,11 @@
  * limitations under the License.
  */
 
-import type { Expr } from "../types";
-import type { CypherEnvironment } from "../Environment";
+import * as CypherBuilder from "../src";
 
-export function serializeMap(
-    env: CypherEnvironment,
-    obj: Record<string, Expr | undefined>,
-    omitCurlyBraces = false
-): string {
-    const valuesList = Object.entries(obj)
-        .filter(([, value]) => value !== undefined)
-        .map(([key, value]) => {
-            return `${key}: ${(value as Expr).getCypher(env)}`; // TODO: improve Typings
-        });
-
-    const serializedContent = valuesList.join(", ");
-    if (omitCurlyBraces) return serializedContent;
-
-    return `{ ${serializedContent} }`;
-}
+describe("Cypher.Utils", () => {
+    test("escapeLabel", () => {
+        const label = CypherBuilder.Utils.escapeLabel("TestLabel");
+        expect(label).toBe("`TestLabel`");
+    });
+});
